It seems page-linear is fixed. The problem we had before was we tried to pagedir_set_page the same address twice - once in install_frame() and one in pull_from_frame().
page_linear has 2500+ page faults. Hopefully the clock hand would fix things up a bit.

+++ pintos/src/userprog/exception.c
@@ -212,13 +212,14 @@ page_fault (struct intr_frame *f)
     }
   }
   else {
-//     printf("Invalid Access.\n");
 //     printf("not_present: %d\n", not_present);
 //     printf("write: %d\n", write);
 //     printf("user: %d\n", user);
 //     printf("fault_addr: %x\n", (uint32_t) fault_addr);
-//     if (!load_page(fault_addr))
-    kill_process(f);
+    if (!load_page(fault_addr)) {
+//       printf("Invalid Access.\n");
+      kill_process(f);
+    }
   }
 }
